requirements.txt
blinker==1.9.0
certifi==2025.4.26
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
slack_sdk==3.35.0
Werkzeug==3.1.3
